fix: config.yaml not mounted with docker-compose
fix: added retry/wait logic when DB is unavailable at start

feat: clear log message when DB credentials are wrong

refactor: moved databse logic into separate database folder
