+++ Proxy/internal/cluster/cluster.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"time"
 
-	model "github.com/amosproj/amos2023ss04-kubernetes-inventory-taker/Proxy/internal/model"
+	model "github.com/amosproj/amos2023ss04-kubernetes-inventory-taker/Proxy/internal/database/model"
 	"github.com/uptrace/bun"
 	"github.com/uptrace/bun/schema"
 	"gopkg.in/yaml.v2"
