+++ Proxy/internal/cluster/service_process.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"log"
 
-	model "github.com/amosproj/amos2023ss04-kubernetes-inventory-taker/Proxy/internal/model"
+	model "github.com/amosproj/amos2023ss04-kubernetes-inventory-taker/Proxy/internal/database/model"
 	"github.com/uptrace/bun"
 	"github.com/uptrace/bun/schema"
 	corev1 "k8s.io/api/core/v1"
